==> 01-get_animes.ipynb <==
import requests
import time

responses = []
page = 1
has_next_page = True

while (has_next_page):
    print(page)
    response = requests.get(f'https://api.jikan.moe/v4/anime?order_by=mal_id&page={page}')
    response_json = response.json()
    responses.append(response_json)
    page = page + 1
    has_next_page = response_json['pagination']['has_next_page']
    time.sleep(1.01)


import joblib
joblib.dump(responses, 'anime_responses.pkl')

import sqlite3 as sq
import pandas as pd

db_name = 'animedb.sqlite'
conn = sq.connect(db_name)
df = pd.read_sql('select * from anime', conn)
conn.close()

from itertools import chain
response_list = list(chain(*[r.get('data', []) for r in responses]))

response_list
