==> 03-get-extended-data-anime.ipynb <==
from itertools import chain
import joblib
import requests
import time

response_list = joblib.load('anime_responses_keep.pkl')

len([r for r in response_list if r['mal_id'] > 0]), len(response_list)

import pandas as pd
pd.DataFrame([r['mal_id'] for r in response_list])

staffs = []
for anime in response_list:
    try:
        print(len(staffs))
        anime_id = anime['mal_id']
        response = requests.get(f'https://api.jikan.moe/v4/anime/{anime_id}/staff')
        staffs.append([anime_id, response.json()])
        if (len(staffs)%500) == 0:
            joblib.dump(staffs, f'staffs-{len(staffs)}.pkl')
        time.sleep(1)
    except:
        print(f'Exception: {anime_id}')
        staffs.append([anime_id, None])

characters = []

for anime in response_list:
    try:
        print(len(characters))
        anime_id = anime['mal_id']
        response = requests.get(f'https://api.jikan.moe/v4/anime/{anime_id}/characters')
        characters.append([anime_id, response.json()])
        if (len(characters)%500) == 0:
            joblib.dump(characters, f'characters-{len(characters)}.pkl')
        time.sleep(1)
    except:
        print(f'Exception: {anime_id}')
        staffs.append([anime_id, dict()])

(
    len([s for s in staffs if 'data' in s[1]]),
    len([s for s in characters if 'data' in s[1]])
)

joblib.dump(staffs, 'staffs.pkl')
joblib.dump(characters, 'characters.pkl')
