==> 04-create-db.ipynb <==
import sqlite3 as sq

with open('dbini.sql', 'r') as sql_file:
    sql_script = sql_file.read()

db = sq.connect('animedb6.sqlite')
cursor = db.cursor()
cursor.executescript(sql_script)
db.commit()
db.close()
