==> 04-insert_into_db_characters.ipynb <==
import joblib
import pandas as pd
import sqlite3 as sq

from itertools import chain

from src.db import (
    get_schemas,
    update_db,
    get_tables,
    count_tables
)

from src.format import format_all_characters

db = 'animedb.sqlite'

characters = joblib.load('characters.pkl')

get_schemas(db = db)

tables = get_tables(db = db)
count_tables(tables, db = db)

dict_df_characters = format_all_characters(characters)

df_character = dict_df_characters['df_character']
df_voiceactors = dict_df_characters['df_voiceactors']
df_character_voiceactors = dict_df_characters['df_character_voiceactors']
df_anime_character = dict_df_characters['df_anime_character']

update_db(df_character,             'character'           , ['character_id'],                  db=db)
update_db(df_voiceactors,           'voiceactor'          , 'voiceactor_id',                   db=db, recreate=False)
update_db(df_character_voiceactors, 'character_voiceactor', ['character_id', 'voiceactor_id'], db=db)
update_db(df_anime_character,       'anime_character'     , ['anime_id', 'character_id'],      db=db)

count_tables(tables, db = db)
