==> 12-graph-test.ipynb <==
import json

# anime
#   character
#   staff
# character
#   voiceactor
#   anime
# voiceactor
#   character
# staff
#   anime

with open('animu_erdos/src/animedbv2.json', 'r') as f:
    db_list = json.load(f)

nodes = [
  f'{d["itemType"]}_{d["id"]}'
  for d in db_list
]

relations = [
  (f'{d_from["itemType"]}_{d_from["id"]}', f'{d_to[1]}_{d_to[0]}')
  for d_from in db_list
  for d_to in d_from.get('relations',[])
]

nodes

import pandas as pd
import networkx as nx

G = nx.Graph()
G.add_nodes_from(nodes)
G.add_edges_from(relations)

nx.has_path(G, 'anime_16498', 'character_1')

def df_to_node_attribute(df, exclude_cols, dict_constants):
    return df\
        .drop(columns=exclude_cols)\
        .set_index('id')\
        .assign(**dict_constants)\
        .to_dict(orient = 'index')

# node_attributes = {
#     **df_to_node_attribute(df_anime,      'anime_id',      dict(nodetype='anime')),
#     **df_to_node_attribute(df_character,  'character_id',  dict(nodetype='character')),
#     **df_to_node_attribute(df_voiceactor, 'voiceactor_id', dict(nodetype='voiceactor')),
#     **df_to_node_attribute(df_staff,      'staff_id',      dict(nodetype='staff')),
# }

# nx.set_node_attributes(G, node_attributes)

nx.write_graphml(G, 'web/animegraph.graphml')

nx.write_gexf(G, 'web/animegraph.gexf')

import json
with open('web/cytograph.json', 'w') as f:
    json.dump(nx.cytoscape_data(G), f)

from src.api import request_api_get

r = await request_api_get('users/basemp44/animelist')

r
